add subbrands when appropiate
