+++ applications/fishing-map/src/features/sidebar/SidebarHeader.module.css
@@ -1,0 +1,25 @@
+:global(.sticky) {
+  background-color: var(--color-white);
+  box-shadow: var(--box-shadow);
+  z-index: 1;
+}
+
+.sidebarHeader {
+  display: flex;
+  align-items: center;
+  padding: 0 var(--space-S);
+  border-bottom: var(--border);
+  height: 5.2rem;
+}
+
+.logo {
+  width: 18rem;
+  height: auto;
+  margin-right: auto;
+}
+
+@media print {
+  .logo {
+    margin-left: var(--space-S);
+  }
+}
